+++ src/mesh/geometry/mesh-geometry-attribute.ts
@@ -23,5 +23,9 @@ export interface MeshGeometryAttribute {
      * is tightly packed. When two or more attributes use the same buffer, this
      * field must be defined.
      */
-     stride?: number
+    stride?: number
+    /**
+     * The number of elements in this attribute.
+     */
+    componentCount?: number
 }